+++ pennylane_lightning/src/gates/cpu_kernels/GateImplementationsLM.hpp
@@ -16,6 +16,7 @@
  * Defines kernel functions with less memory (and fast)
  */
 #pragma once
+
 #include "BitUtil.hpp"
 #include "Error.hpp"
 #include "GateOperation.hpp"