+++ pennylane_lightning/src/util/ConstantUtil.hpp
@@ -16,6 +16,10 @@
  * Contains utility functions for processing constants
  */
 #pragma once
+
+#include "TypeTraits.hpp"
+#include "Util.hpp"
+
 #include <algorithm>
 #include <array>
 #include <compare>
@@ -26,8 +30,6 @@
 #if __has_include(<version>)
 #include <version>
 #endif
-
-#include "Util.hpp"
 
 namespace Pennylane::Util {
 /**